Create core.magics.config according to new API.
