+++ IPython/core/magics/__init__.py
@@ -15,6 +15,7 @@
 from IPython.core.magic import Magics, register_magics
 from .basic import BasicMagics
 from .code import CodeMagics, MacroToEdit
+from .config import ConfigMagics
 from .history import HistoryMagics
 
 #-----------------------------------------------------------------------------
